separate whitelist from user component
